requirements.txt
aiohttp==3.7.4.post0
anyio==3.4.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
cenzopapa-sdk==1.1.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
decorators==2.0.7
disnake==2.3.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
multidict==5.2.0
python-dotenv==0.19.2
pytz==2021.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
